Fix skymaps not loading in headless mode without JavaFX.
