+++ chunky/src/java/se/llbit/util/ImageTools.java
@@ -16,9 +16,6 @@
  */
 package se.llbit.util;
 
-import javafx.scene.image.Image;
-import javafx.scene.image.PixelFormat;
-import javafx.scene.image.WritableImage;
 import se.llbit.chunky.resources.BitmapImage;
 import se.llbit.math.ColorUtil;
 
@@ -57,13 +54,5 @@ public class ImageTools {
     } else {
       return ColorUtil.getArgb(ra / aa, ga / aa, ba / aa, aa / n);
     }
-  }
-
-  /** @return a JavaFX version of a BitmapImage. */
-  public static Image toFxImage(BitmapImage image) {
-    WritableImage fxImage = new WritableImage(image.width, image.height);
-    fxImage.getPixelWriter().setPixels(0, 0, image.width, image.height,
-        PixelFormat.getIntArgbInstance(), image.data, 0, image.width);
-    return fxImage;
   }
 }
